Fix endpoint location
