+++ hub/backend/app/api/routes/projects/core.py
@@ -2898,7 +2898,6 @@ def _make_comment_title(comment: str) -> str:
     Strips trailing ``.`` and ``!`` but preserves ``?`` so questions read
     naturally as titles.
     """
-    import re
 
     m = re.search(r"([.!?])\s", comment)
     if m:
@@ -5915,167 +5914,6 @@ def delete_project_references(
         add_event_info={"path": path},
     )
     return Message(message="References collection deleted")
-
-
-def _normalize_doi(value: str | None) -> str | None:
-    """Reduce a DOI to a bare, comparable form, e.g. ``10.1234/abcd``."""
-    if not value:
-        return None
-    doi = value.strip().lower()
-    for prefix in ("https://doi.org/", "http://doi.org/", "doi:"):
-        if doi.startswith(prefix):
-            doi = doi[len(prefix) :]
-    return doi.strip("/ ") or None
-
-
-def _normalize_arxiv_id(value: str | None) -> str | None:
-    """Reduce an arXiv ID to a bare, version-less form, e.g. ``2301.01234``."""
-    if not value:
-        return None
-    arxiv_id = value.strip().lower()
-    for prefix in (
-        "https://arxiv.org/abs/",
-        "http://arxiv.org/abs/",
-        "arxiv:",
-    ):
-        if arxiv_id.startswith(prefix):
-            arxiv_id = arxiv_id[len(prefix) :]
-    arxiv_id = re.sub(r"v\d+$", "", arxiv_id.strip("/ "))
-    return arxiv_id or None
-
-
-def _normalize_title(value: str | None) -> str | None:
-    """Reduce a title to letters and digits so punctuation, case, and LaTeX
-    braces don't defeat a comparison.
-    """
-    if not value:
-        return None
-    return re.sub(r"[^a-z0-9]+", "", value.lower()) or None
-
-
-class ReferenceSearchMatch(BaseModel):
-    project_owner_name: str
-    project_name: str
-    project_title: str
-    path: str
-    key: str
-    type: str
-    title: str | None = None
-    doi: str | None = None
-    note_count: int = 0
-    matched_on: Literal["doi", "arxiv_id", "title"]
-
-
-@router.get("/user/references/search")
-def get_user_reference_matches(
-    current_user: CurrentUser,
-    session: SessionDep,
-    projects: Annotated[list[str], Query()],
-    doi: str | None = None,
-    arxiv_id: str | None = None,
-    title: str | None = None,
-) -> list[ReferenceSearchMatch]:
-    """Find a reference in the collections of the given projects.
-
-    Each entry in ``projects`` is an ``owner/name`` pair. The projects are
-    named explicitly rather than searched across everything the user can
-    read, since each one has to be cloned and read, which would make an
-    unbounded search unusably slow.
-    """
-    target_doi = _normalize_doi(doi)
-    target_arxiv_id = _normalize_arxiv_id(arxiv_id)
-    target_title = _normalize_title(title)
-    if not any([target_doi, target_arxiv_id, target_title]):
-        raise HTTPException(422, "A DOI, arXiv ID, or title is required")
-    if not projects:
-        raise HTTPException(422, "At least one project is required")
-    if len(projects) > 10:
-        raise HTTPException(422, "At most 10 projects can be searched")
-    resp = []
-    for project_spec in projects:
-        if project_spec.count("/") != 1:
-            raise HTTPException(
-                422, f"Invalid project '{project_spec}'; expected owner/name"
-            )
-        spec_owner_name, spec_project_name = project_spec.split("/")
-        try:
-            project = app.projects.get_project(
-                owner_name=spec_owner_name,
-                project_name=spec_project_name,
-                session=session,
-                current_user=current_user,
-                min_access_level="read",
-            )
-            repo = get_repo(
-                project=project,
-                user=current_user,
-                session=session,
-                ttl=DEFAULT_REPO_TTL,
-            )
-        except HTTPException as e:
-            # One inaccessible project shouldn't sink the whole lookup
-            logger.info(f"Skipping project '{project_spec}': {e.detail}")
-            continue
-        ck_info = get_ck_info_from_repo(repo)
-        bib_paths = {
-            rc["path"]
-            for rc in (ck_info.get("references") or [])
-            if isinstance(rc, dict) and "path" in rc
-        }
-        try:
-            for blob in repo.head.commit.tree.traverse():
-                if blob.type != "blob":  # type: ignore[union-attr]
-                    continue
-                blob_path = str(blob.path)  # type: ignore[union-attr]
-                if not blob_path.lower().endswith(".bib"):
-                    continue
-                if any(p.startswith(".") for p in blob_path.split("/")):
-                    continue
-                bib_paths.add(blob_path)
-        except Exception as e:
-            logger.warning(f"Failed to scan for .bib files: {e}")
-        for bib_path in sorted(bib_paths):
-            full_path = os.path.join(repo.working_dir, bib_path)
-            if not os.path.isfile(full_path):
-                continue
-            try:
-                with open(full_path) as f:
-                    bib_db = bibtexparser.loads(f.read())
-            except Exception as e:
-                logger.warning(f"Failed to parse BibTeX file {bib_path}: {e}")
-                continue
-            for entry in bib_db.entries:
-                entry_doi = _normalize_doi(entry.get("doi"))
-                entry_arxiv_id = _normalize_arxiv_id(
-                    arxiv.id_from_bib_attrs(entry)
-                )
-                entry_title = _normalize_title(entry.get("title"))
-                if target_doi and entry_doi == target_doi:
-                    matched_on = "doi"
-                elif target_arxiv_id and entry_arxiv_id == target_arxiv_id:
-                    matched_on = "arxiv_id"
-                elif target_title and entry_title == target_title:
-                    matched_on = "title"
-                else:
-                    continue
-                comment = entry.get(BIB_NOTE_FIELD)
-                resp.append(
-                    ReferenceSearchMatch(
-                        project_owner_name=project.owner_account_name,
-                        project_name=project.name,
-                        project_title=project.title,
-                        path=bib_path,
-                        key=entry.get("ID", ""),
-                        type=entry.get("ENTRYTYPE", "misc"),
-                        title=entry.get("title"),
-                        doi=entry.get("doi"),
-                        note_count=len(
-                            zotero.parse_notes_markdown(comment or "")
-                        ),
-                        matched_on=matched_on,  # type: ignore[arg-type]
-                    )
-                )
-    return resp
 
 
 class ReferenceItemPost(BaseModel):
