+++ hub/frontend/src/client/index.ts
@@ -490,11 +490,11 @@ export type {
   GetUserOverleafTokenData,
   GetUserOverleafTokenResponse,
   GetUserOverleafTokenResponses,
-  GetUserReferenceMatchesData,
-  GetUserReferenceMatchesError,
-  GetUserReferenceMatchesErrors,
-  GetUserReferenceMatchesResponse,
-  GetUserReferenceMatchesResponses,
+  GetUserReferencesData,
+  GetUserReferencesError,
+  GetUserReferencesErrors,
+  GetUserReferencesResponse,
+  GetUserReferencesResponses,
   GetUserStorageData,
   GetUserStorageResponse,
   GetUserStorageResponses,
