+++ hub/frontend/src/client/sdk.gen.ts
@@ -200,8 +200,8 @@ import type {
   GetUserOverleafSyncErrors,
   GetUserOverleafSyncResponses,
   GetUserOverleafTokenResponses,
-  GetUserReferenceMatchesErrors,
-  GetUserReferenceMatchesResponses,
+  GetUserReferencesErrors,
+  GetUserReferencesResponses,
   GetUserStorageResponses,
   GetUserTokensErrors,
   GetUserTokensResponses,
@@ -1828,6 +1828,64 @@ export class UsersService {
       security: [{ scheme: "bearer", type: "http" }],
       url: "/user/storage",
       ...options,
+    })
+  }
+
+  /**
+   * Get User References
+   *
+   * Read the reference collections across the user's projects.
+   *
+   * Every filter is optional. With none, this lists every entry it finds;
+   * with ``doi``, ``arxiv_id``, or ``title``, only entries matching one of
+   * them, which is how a client asks "is this paper already filed?".
+   *
+   * ``projects`` narrows the search to the given ``owner/name`` pairs.
+   * Without it, the search covers the projects the user can write to, which
+   * is what "their" references means -- read access would pull in every
+   * public project on the hub.
+   *
+   * Each project has to be cloned and read, so ``max_projects`` bounds the
+   * work. Projects beyond it are not searched.
+   */
+  public static getUserReferences<ThrowOnError extends boolean = true>(
+    parameters?: {
+      projects?: Array<string> | null
+      doi?: string | null
+      arxiv_id?: string | null
+      title?: string | null
+      max_projects?: number
+    },
+    options?: Options<never, ThrowOnError>,
+  ): RequestResult<
+    GetUserReferencesResponses,
+    GetUserReferencesErrors,
+    ThrowOnError
+  > {
+    const params = buildClientParams(
+      [parameters],
+      [
+        {
+          args: [
+            { in: "query", key: "projects" },
+            { in: "query", key: "doi" },
+            { in: "query", key: "arxiv_id" },
+            { in: "query", key: "title" },
+            { in: "query", key: "max_projects" },
+          ],
+        },
+      ],
+    )
+    return (options?.client ?? client).get<
+      GetUserReferencesResponses,
+      GetUserReferencesErrors,
+      ThrowOnError
+    >({
+      responseType: "json",
+      security: [{ scheme: "bearer", type: "http" }],
+      url: "/user/references",
+      ...options,
+      ...params,
     })
   }
 }
@@ -4877,55 +4935,6 @@ export class ProjectsService {
         ...options?.headers,
         ...params.headers,
       },
-    })
-  }
-
-  /**
-   * Get User Reference Matches
-   *
-   * Find a reference in the collections of the given projects.
-   *
-   * Each entry in ``projects`` is an ``owner/name`` pair. The projects are
-   * named explicitly rather than searched across everything the user can
-   * read, since each one has to be cloned and read, which would make an
-   * unbounded search unusably slow.
-   */
-  public static getUserReferenceMatches<ThrowOnError extends boolean = true>(
-    parameters: {
-      projects: Array<string>
-      doi?: string | null
-      arxiv_id?: string | null
-      title?: string | null
-    },
-    options?: Options<never, ThrowOnError>,
-  ): RequestResult<
-    GetUserReferenceMatchesResponses,
-    GetUserReferenceMatchesErrors,
-    ThrowOnError
-  > {
-    const params = buildClientParams(
-      [parameters],
-      [
-        {
-          args: [
-            { in: "query", key: "projects" },
-            { in: "query", key: "doi" },
-            { in: "query", key: "arxiv_id" },
-            { in: "query", key: "title" },
-          ],
-        },
-      ],
-    )
-    return (options?.client ?? client).get<
-      GetUserReferenceMatchesResponses,
-      GetUserReferenceMatchesErrors,
-      ThrowOnError
-    >({
-      responseType: "json",
-      security: [{ scheme: "bearer", type: "http" }],
-      url: "/user/references/search",
-      ...options,
-      ...params,
     })
   }
